[MOV] website_slides, *: move widgets to interactions folder
*: website_sale_slides, website_slides_survey

task-5353853
