+++ addons/website_sale_slides/static/src/interactions/slides_course_join.js
@@ -1,4 +1,4 @@
-import CourseJoin from "@website_slides/js/slides_course_join";
+import CourseJoin from "@website_slides/interactions/slides_course_join";
 
 const CourseJoinWidget = CourseJoin.courseJoinWidget;
