+++ addons/website_slides/__manifest__.py
@@ -83,12 +83,6 @@ Featuring
             'website_slides/static/src/scss/website_slides.scss',
             'website_slides/static/src/scss/website_slides_profile.scss',
             'website_slides/static/src/scss/slides_slide_fullscreen.scss',
-            'website_slides/static/src/js/slides_course_page.js',
-            'website_slides/static/src/js/slides_course_slides_list.js',
-            'website_slides/static/src/js/slides_course_fullscreen_player.js',
-            'website_slides/static/src/js/slides_course_join.js',
-            'website_slides/static/src/js/slides_course_quiz.js',
-            'website_slides/static/src/js/slides_course_quiz_question_form.js',
             'website_slides/static/src/xml/website_slides_sidebar.xml',
             'website_slides/static/src/xml/website_slides_fullscreen.xml',
             'website_slides/static/src/xml/slide_management.xml',
