+++ addons/website_slides/static/src/interactions/slides_course_join.js
