+++ addons/website_slides/static/src/interactions/slides_course_quiz_question_form.js
