+++ addons/website_slides/static/src/interactions/slides_course_slides_list.js
@@ -1,6 +1,6 @@
 import publicWidget from '@web/legacy/js/public/public_widget';
 import { _t } from "@web/core/l10n/translation";
-import { SlideCoursePage } from '@website_slides/js/slides_course_page';
+import { SlideCoursePage } from '@website_slides/interactions/slides_course_page';
 
 publicWidget.registry.websiteSlidesCourseSlidesList = SlideCoursePage.extend({
     selector: '.o_wslides_slides_list',
