+++ addons/website_slides/static/tests/tours/slides_course_member_yt.js
@@ -6,7 +6,7 @@ function patchFullScreen() {
      * This will make the video start at 10 minutes.
      * As it lasts 10min24s, it will mark it as completed immediately.
     */
-    const FullScreen = odoo.loader.modules.get('@website_slides/js/slides_course_fullscreen_player')[Symbol.for("default")];
+    const FullScreen = odoo.loader.modules.get('@website_slides/interactions/slides_course_fullscreen_player')[Symbol.for("default")];
     FullScreen.include({
         _renderSlide: function () {
