+++ addons/website_slides_survey/__manifest__.py
@@ -34,7 +34,7 @@
     'assets': {
         'web.assets_frontend': [
             'website_slides_survey/static/src/scss/website_slides_survey.scss',
-            'website_slides_survey/static/src/js/slides_course_fullscreen_player.js',
+            'website_slides_survey/static/src/interactions/*.js',
             'website_slides_survey/static/src/xml/website_slides_fullscreen.xml',
             'website_slides_survey/static/src/js/public/**/*',
         ],
