+++ addons/website_slides_survey/static/src/interactions/slides_course_fullscreen_player.js
@@ -1,5 +1,5 @@
 import { renderToElement } from "@web/core/utils/render";
-import Fullscreen from "@website_slides/js/slides_course_fullscreen_player";
+import Fullscreen from "@website_slides/interactions/slides_course_fullscreen_player";
 
 Fullscreen.include({
     /**
